Trees, Wind and Tweaks
Exchanged old trees for default Tree Generator model
Added a wind zone. Isn't it so peaceful.
Made the entire house the loot hoard.
Added the lazrbike to one of the garages.
Organized the files in the main directory, I don't think I deleted anything important.
Organized the items in the game scene.